src/ai_service/requirements.txt
flask>=2.3.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0
